pyproject.toml
[project]
name = "responses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.1.0",
]
